Fixed some minor remarks
