+++ src/app/shared/autocomplete/city.interface.ts
@@ -1,0 +1,6 @@
+export interface City {
+  country: string;
+  geonameid: number;
+  name: string;
+  subcountry: string;
+}
