+++ src/app/shared/constants/autocomplete-form.constants.ts
@@ -1,0 +1,1 @@
+export const AGE_OPTIONS = ['Under 18', '18-25', '25-40', 'Over 40'];
