+++ src/app/variables.less
@@ -1,2 +1,3 @@
 @primary-red-color: #F40009;
 @white_deep: #fff;
+@lightgray: #CFCDCC;
